+++ tests_load.py
@@ -1,4 +1,5 @@
 import json
+import os
 
 question = []
 answer = []
@@ -22,7 +23,33 @@ quest = {"quests":question,
 
 to_json = {'test': quest}
 
-with open('test/tests1.json', 'w') as f:
-    json.dump(to_json, f,sort_keys=True, indent=3,ensure_ascii=False )
+
+# with open('tests1.json') as f:
+#     data = json.load(f)
+#
+#     data['test']['quests'].extend(question)
+#     data['test']['answer'].extend(answer)
+#     data['test']['true_answer'].extend(true_answer)
+#     print(data)
+#
+#
+#
+# with open('tests1.json', 'w') as f:
+#     json.dump(data, f,sort_keys=True, indent=3,ensure_ascii=False )
+
+if os.path.isfile('tests1.json') and os.access('tests1.json',os.R_OK):
+    with open('tests1.json') as f:
+        data = json.load(f)
+
+    data['test']['quests'].extend(question)
+    data['test']['answer'].extend(answer)
+    data['test']['true_answer'].extend(true_answer)
+    print(data)
+
+    with open('tests1.json', 'w') as f:
+        json.dump(data, f,sort_keys=True, indent=3,ensure_ascii=False )
+else:
+    with open('tests1.json', 'w') as f:
+        json.dump(to_json, f,sort_keys=True, indent=3,ensure_ascii=False )
